redid structure and testframe
