+++ src/test/test_master.py
@@ -6,7 +6,3 @@ def master_suite():
     suite.addTest(MidiParserTestCase('test_validate'))
     suite.addTest(MidiParserTestCase('test_parse'))
     return suite
-
-if __name__ == '__main__':
-    runner = unittest.TextTestRunner()
-    runner.run(master_suite())
