--- test/test_midi_parser.py
@@ -1,14 +1,0 @@
-import unittest
-import sys
-sys.path.insert(0, '../src/midiParser')
-from MidiParser import *
-LOG_HEADER ="\n[MidiParserTestCase]"
-class MidiParserTestCase(unittest.TestCase):
-
-
-    def test_validate(self):
-        print(sys.path)
-        self.fail('Not implemented yet.')
-    def test_parse(self):
-        self.fail('Not implemented yet.')
-
